handled the return points string to add the customer's name

+++ src/main/java/com/portfolio/melkhazen/chartercommunications/entity/Customer.java
@@ -19,5 +19,5 @@ public class Customer {
     @Column(name = "TRANSACTION_DATE")
     private String transactionDate;
     @Column(name = "AMOUNT")
-    private double transactionAmount;
+    private int transactionAmount;
 }
